requirements.txt
mysql-connector-python
pandas
python-dotenv
matplotlib
seaborn
SQLAlchemy
pyarrow
scikit-learn
imbalanced-learn
fastparquet
joblib
scipy
xgboost
numpy